refactor: Valid를 좀더 구체적으로 변경 및 일부 검증은 해당 클래스로 이동
