+++ src/main/java/racing/utils/ParserUtils.java
@@ -1,13 +1,11 @@
 package racing.utils;
 
-import java.util.Optional;
-
 public class ParserUtils {
 
     private static final String COMMA = ",";
 
     public static String[] stringToArray(String input) {
-        Valid.isEmpty(input);
+        InputValid.isEmpty(input);
         return input.split(COMMA);
 
     }
